* Update .gitignore with standard VS rules

* Organize solution items in .sln file

* Add support for .NET 10.0

* Bump dependencies

* Migrate to xUnit v3 and adapt test types

* Implement Error equality and update test cases for current API behavior

* Update CreateCheckTest data to USDT and improve helper

* Update GetInvoicesTest data to match actual API errors

* Update TransferTest data & GetTransfersTest data with USDT and to match actual API errors

* Improve test documentation and setup requirements

* Add optional Message property to Error model for detailed API error descriptions

- This field is populated only in rare cases when the API provides additional context about the error. Including it in RequestException improves debugging and provides better insight into specific API responses.

* Change currency types to decimal across the solution

+++ CryptoPay/Types/Balance.cs
@@ -17,7 +17,7 @@ public sealed class Balance
     /// Number of available coins.
     /// </summary>
     [JsonRequired]
-    public double Available { get; set; }
+    public decimal Available { get; set; }
 
     /// <summary>
     /// Unavailable amount currently is on hold in float.
